Authorization consolidation
Auth, buttons
